+++ tests/test_cli.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import unittest.mock as mock
 import subprocess
@@ -5,6 +6,7 @@ import sys
 
 import pytest
 
+import staticjinja
 from staticjinja import cli
 
 srcpath_cases = [
@@ -19,7 +21,7 @@ srcpath_cases = [
 @pytest.mark.parametrize("srcpath, expected", srcpath_cases)
 @mock.patch("os.path.isdir")
 @mock.patch("os.getcwd")
-@mock.patch("staticjinja.cli.Site.make_site")
+@mock.patch("staticjinja.cli.staticjinja.Site.make_site")
 def test_srcpath(mock_make_site, mock_getcwd, mock_isdir, srcpath, expected):
     """Test that various `--srcpath` args given to the CLI result in
     Site.make_site() being called with the correct searchpath parameter."""
@@ -48,7 +50,7 @@ outpath_cases = [
 @pytest.mark.parametrize("outpath, expected", outpath_cases)
 @mock.patch("os.path.isdir")
 @mock.patch("os.getcwd")
-@mock.patch("staticjinja.cli.Site.make_site")
+@mock.patch("staticjinja.cli.staticjinja.Site.make_site")
 def test_outpath(mock_make_site, mock_getcwd, mock_isdir, outpath, expected):
     """Test that various `--outpath` args given to the CLI result in
     Site.make_site() being called with the correct outpath parameter."""
@@ -65,6 +67,28 @@ def test_outpath(mock_make_site, mock_getcwd, mock_isdir, outpath, expected):
     )
 
 
+@mock.patch("os.path.isdir")
+@mock.patch("os.getcwd")
+def test_log(mock_getcwd, mock_isdir):
+    mock_isdir.return_value = True
+    mock_getcwd.return_value = "/cwd"
+
+    # Passing arg sets logger level
+    argv = ["build", "--log=critical"]
+    cli.main(argv)
+    assert staticjinja.logger.level == logging.CRITICAL
+
+    # Default log level is INFO if not set
+    argv = ["build"]
+    cli.main(argv)
+    assert staticjinja.logger.level == logging.INFO
+
+    # Bogus level is caught
+    with pytest.raises(ValueError):
+        argv = ["build", "--log=junk"]
+        cli.main(argv)
+
+
 @pytest.mark.parametrize(
     "command, expected",
     [
@@ -74,7 +98,7 @@ def test_outpath(mock_make_site, mock_getcwd, mock_isdir, outpath, expected):
 )
 @mock.patch("os.path.isdir")
 @mock.patch("os.getcwd")
-@mock.patch("staticjinja.cli.Site.make_site")
+@mock.patch("staticjinja.cli.staticjinja.Site.make_site")
 def test_watch(mock_make_site, mock_getcwd, mock_isdir, command, expected):
     """Test that build/watch commands result in Site.render() being called
     with the correct use_reloader values."""
@@ -88,7 +112,7 @@ def test_watch(mock_make_site, mock_getcwd, mock_isdir, command, expected):
     mock_site.render.assert_called_once_with(use_reloader=expected)
 
 
-@mock.patch("staticjinja.cli.Site.make_site")
+@mock.patch("staticjinja.cli.staticjinja.Site.make_site")
 def test_nonexistent_srcpath(mock_make_site):
     """Test that a nonexistent `--srcpath` exits early."""
     # Technique from